requirements9.txt
agora-token-builder==1.0.0
asgiref==3.5.0
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
beautifulsoup4==4.11.1
blis==0.7.8
catalogue==2.0.7
certifi==2022.6.15
cffi==1.15.0
channels==3.0.5
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.2
cymem==2.0.6
daphne==3.0.2
Django==3.2
django-bootstrap4==22.1
django-ckeditor==6.4.0
django-crispy-forms==1.11.2
django-embed-video==1.4.4
django-filter==22.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-mailjet==0.3.1
django-recaptcha==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
filelock==3.7.1
flashtext==2.7
future==0.18.2
huggingface-hub==0.8.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
install-jdk==0.3.0
javac-parser==1.0.0
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
mailjet-rest==1.3.4
MarkupSafe==2.1.1
msgpack-python==0.5.6
murmurhash==1.0.7
networkx==2.8.4
nltk==3.7
numpy==1.23.1
packaging==21.3
pathy==0.6.2
Pillow==9.1.0
pke==2.0.0
preshed==3.0.6
py4j==0.10.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
sentence-transformers==2.2.2
sentencepiece==0.1.96
service-identity==21.1.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
sqlparse==0.4.2
srsly==2.4.3
thinc==8.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
transformers==4.20.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typer==0.4.2
typing_extensions==4.2.0
Unidecode==1.3.4
urllib3==1.26.9
wasabi==0.9.1
zope.interface==5.4.0
